systemd: removed enlightenment-user.path
enlightenment-user.service wil be activated when starting basic.target.wants
and check itself whether /run/wayland-0 socket file exists using shell scripts
in ExecStart.

If the path-based activation is used, then enlightenment-user.service is not
registered as a job of the systemd. In this case, the systemd could ignore
dependencies on the enlightenment-user.service.

Change-Id: Iee9ddd39f4c2dca1c2a444d916e226d289c31ae1